Add files via upload
A detailed analysis for an insurance company has been performed to reduce the risk and to increase profit for the organization. The analysis comprises of different regression models which are measured using the performance metrics such as RMSE, MSE and MAE. The proposed report achieved to estimate the insurance cost at 95% accuracy and provided recommendations to reduce the risk and to maximize profit.